Refactor simple-autopair and introduce smart space key
Since I use simple-autopair only for enh-ruby-mode, the
'simple-autopair-configure function is now gone.

Now, all keys are defined into the minor mode map instead of being
dynamically hooked with local-set-key into the major mode.

The new 'simple-autopair-enabled-pairs variable defines which pairs
are enabled for expansion. This isn't strictly necessary for now, but
if I ever use this library with another mode in the future, I will be
able to override 'simple-autopair-enabled-pairs with setq-local. Also,
I'd need to figure out how to make simple-autopair aware of string
delimiters regardless of major mode (currently it only works for
enh-ruby-mode,) or allow extending this behavior per mode with custom
logic.

Now onto the smart space key. Imagine the cursor to be the pipe below.
The smart space key is great because if I type "{", it will get
expanded into:

{|}

Now, by pressing space, I get:

{ | }

This behavior was previously automatic. "{" would be directly expanded
into "{ }". But many times, all I want is an empty hash, so it
shouldn't be spaced out automatically. And additional "space" will
space out the "{}" pair, or whatever pairs I configure with the
'simple-autopair-spaced variable.